[AS7-2733] Clean up hack. All action can take place in the domain operation handlers now that we can see the server inventory
